pyproject.toml
[tool.poetry]
name = "open-ai-search"
version = "0.1.0"
description = "Open AI Search"
authors = ["Lucien Shui <lucien@lucien.ink>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
openai = "^1.57.4"
sse-starlette = "^2.1.3"
pyyaml = "^6.0.2"
duckduckgo-search = "^6.4.1"
shortuuid = "^1.0.13"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"


[tool.poetry.group.backport.dependencies]
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
bilibili-api-python = "^16.3.0"
python-dateutil = "^2.9.0.post0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
